Delete old swap module
