+++ app/src/main/java/io/horizontalsystems/bankwallet/modules/balance/BalanceFragment.kt
@@ -25,7 +25,7 @@ import io.horizontalsystems.bankwallet.modules.ratechart.RateChartActivity
 import io.horizontalsystems.bankwallet.modules.receive.ReceiveFragment
 import io.horizontalsystems.bankwallet.modules.settings.managekeys.views.ManageKeysDialog
 import io.horizontalsystems.bankwallet.modules.settings.security.privacy.PrivacySettingsModule
-import io.horizontalsystems.bankwallet.modules.swapnew.SwapModuleNew
+import io.horizontalsystems.bankwallet.modules.swap.SwapModule
 import io.horizontalsystems.bankwallet.ui.extensions.NpaLinearLayoutManager
 import io.horizontalsystems.bankwallet.ui.extensions.SelectorDialog
 import io.horizontalsystems.bankwallet.ui.extensions.SelectorItem
@@ -178,7 +178,7 @@ class BalanceFragment : Fragment(), BalanceItemsAdapter.Listener, ReceiveFragmen
         })
 
         viewModel.openSwap.observe(viewLifecycleOwner, Observer { wallet ->
-            context?.let { context -> SwapModuleNew.start(context, wallet.coin) }
+            context?.let { context -> SwapModule.start(context, wallet.coin) }
         })
 
         viewModel.didRefreshLiveEvent.observe(viewLifecycleOwner, Observer {
