+++ app/src/main/java/io/horizontalsystems/bankwallet/modules/swap/model/AmountType.kt
@@ -1,0 +1,5 @@
+package io.horizontalsystems.bankwallet.modules.swap.model
+
+enum class AmountType {
+    ExactSending, ExactReceiving
+}
