--- app/src/main/java/io/horizontalsystems/bankwallet/modules/swapnew/model/AmountType.kt
@@ -1,5 +1,0 @@
-package io.horizontalsystems.bankwallet.modules.swapnew.model
-
-enum class AmountType {
-    ExactSending, ExactReceiving
-}
